More error handler definition to server init
