+++ StarcorpServer/starcorp/server/__init__.py
@@ -1,16 +1,18 @@
 """ Package for containing server setup and logic. """
 # pylint: disable=wrong-import-position
 
-import os
 import logging
+import os
 from logging.handlers import RotatingFileHandler
+from uuid import uuid4
 
 import flask
 from flask import Flask
 from flask_socketio import SocketIO
 
+from data import CONFIG, TYPE_META, Decoder, Encoder, Serializable
+from exceptions import HttpError, SocketIOEventError
 from utils.logging import FORMATTER
-from data import TYPE_META, Decoder, Encoder, Serializable, CONFIG
 
 socketio = SocketIO()
 
@@ -38,6 +40,29 @@ def convert_custom_object(obj):
 
 
 app.make_response = convert_custom_object
+
+
+@app.errorhandler(HttpError)
+def handle_http_error(error):
+    """ Translate the unhandled error into an appropriate api response. """
+    return error.response
+
+
+@socketio.on_error()
+def handle_error(error):
+    """ Handle generic SocketIOEventError. """
+
+    if isinstance(error, SocketIOEventError):
+        LOGGER.warning(f"Error encountered during {request.event}: {error!r}")
+        emit(error.event, error.response)
+    else:
+        error_id = uuid4()
+        LOGGER.exception(
+            f"An unexpected {error!r} has ocurred during {request.event}. Error id: {error_id}!"
+        )
+        emit("error", {"reason": "An unexpected error ocurred", "id": str(error_id)})
+        disconnect()
+        raise error
 
 
 LOGGER = logging.getLogger("socketio")
